pyproject.toml
[project]
name = "ember"
version = "0.1.0"
description = "Orange Pi voice assistant with wake word detection and ChatGPT integration"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "pyaudio",
    "speechrecognition",
    "requests",
    "openai",
    "pyttsx3",
    "pvporcupine",
    "numpy >1.24.3",
    "grpcio",
    "grpcio-tools",
    "protobuf",
    "vosk>=0.3.45",
    "python-dotenv>=1.0.1",
    "pip>=25.0.1",
    "ipython>=8.12.3",
]
